offstore/requirements.txt
Django
django-debug-toolbar
djangorestframework
djangorestframework_simplejwt
django-cors-headers
